requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.3.2
alembic==1.15.2
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cx_Oracle==8.3.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==2.3.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.4.1
greenlet==3.2.0
gunicorn==23.0.0
h11==0.14.0
http_ece==1.2.1
idna==3.3
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
lxml==5.3.2
Mako==1.1.5
MarkupSafe==3.0.2
mimesis==5.1.1
msgspec==0.19.0
multidict==6.0.5
mysql-connector-python==8.2.0
narwhals==1.34.0
numpy==1.26.4
openpyxl==3.1.5
oracledb==3.1.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
Pillow==10.0.1
plotly==6.0.1
plyer==2.1.0
protobuf==4.21.12
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.5.0
pycparser==2.21
PyMySQL==1.1.1
PySocks==1.7.1
pystray==0.19.5
python-dateutil==2.8.2
python-dotenv==1.1.0
python-editor==1.0.4
python-engineio==4.11.2
python-redmine==2.4.0
python-slugify==7.0.0
python-socketio==5.12.1
python-xlib==0.33
pytz==2024.1
pywebpush==1.14.0
requests==2.32.3
beautifulsoup4==4.12.3
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
text-unidecode==1.3
tqdm==4.66.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.2
visitor==0.1.3
websocket-client==1.8.0
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
xmltodict==0.14.2
yarl==1.9.4
zipp==3.21.0
python-json-logger==3.3.0
concurrent-log-handler==0.9.28
